+++ src/components/community/MarkdownEditor/index.ts
@@ -1,0 +1,2 @@
+export { MarkdownEditor } from './MarkdownEditor'
+export type { MarkdownEditorProps } from './MarkdownEditor'
