+++ src/components/community/index.ts
@@ -6,3 +6,4 @@ export * from './ImageUpload'
 export * from './SubmitButton'
 export * from './CancelButton'
 export * from './PostForm'
+export * from './MarkdownEditor'
